x86/resctrl: Move the filesystem portions of resctrl to live in '/fs/'
maillist inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I8T2RT

Reference: https://git.kernel.org/pub/scm/linux/kernel/git/morse/linux.git/log/?h=mpam/snapshot/v6.7-rc2

---------------------------

Move the parts of resctrl that aren't archictecture specific to live
in /fs/.

This lets other architectures implement resctrl too.
